настройка кастомной команды для создания администратора
