+++ config/settings.py
@@ -23,6 +23,12 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
+    "rest_framework",
+    "django_filters",
+    "rest_framework_simplejwt",
+    "django_extensions",
+    "drf_yasg",
+
     "habits",
     "users",
 ]
