--- users/managements/commands/createsuperuser.py
@@ -1,13 +1,0 @@
-from django.core.management import BaseCommand
-
-from users.models import User
-
-
-class Command(BaseCommand):
-    def handle(self, *args, **options):
-        user = User.objects.create(email="testadmin@sky.pro", chat_id=12345)
-        user.is_staff = True
-        user.is_active = True
-        user.is_superuser = True
-        user.set_password("1234")
-        user.save()
